requirements.txt
absl-py==1.4.0
asgiref==3.5.2
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.1.1
ExifRead==3.0.0
filelock==3.11.0
Flask==2.1.1
Flask-Materialize==0.2.4
Flask-WTF==1.1.1
flatbuffers==23.3.3
gast==0.4.0
GDAL==3.3.1
google-api-core==1.34.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-cloud-vision==3.4.2
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.48.2
h5py==3.8.0
httplib2==0.22.0
idna==3.4
image==1.5.33
importlib-metadata==6.4.1
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
keras==2.12.0
libclang==16.0.0
libcst==1.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
metadata-extractor==1.0
ml-dtypes==0.1.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
Pillow==9.2.0
proto-plus==1.22.2
protobuf
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-docx==0.8.11
python-magic==0.4.27
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sqlparse==0.4.2
sympy==1.11.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
typing-inspect==0.9.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.14.1
WTForms==3.0.1
zipp==3.15.0
